requirements.txt
mcp>=1.22.0,<2.0.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
msal>=1.25.0
requests>=2.31.0
pytest>=7.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
